refactor: 폴더 구조 정리
- components 내부에 도메인별 폴더 생성 및 분류
- 일부 컴포넌트 이름 변경 (Markdown, ProblemViewer...)
