+++ frontend/src/components/Simulation/types.ts
